Rename: src/assets 파일 내 개인 페이지 폴더 생성
